added the prerequisite to an employee registration
this commits brings changes to how we reference the scripts in the master layout page, (scripts are loaded based on page requirement) also i have added the other fields to a an employee as discussed. all other fields from the select option are loaded from the db to omiit typo errors.